backend/requirements.txt
asgiref==3.9.0
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==37.4.0
iniconfig==2.1.0
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.9.0
pytesseract==0.3.13
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
